refactor: rename configs package to config
