+++ scripts/dev/twin/get-address-by-twin-id.go
@@ -56,7 +56,7 @@ type TFChainConfig struct {
 	WsProviderURL string
 }
 
-// implement SubstrateConfig for configs.TFChain
+// implement SubstrateConfig for config.TFChain
 func (c *TFChainConfig) GetWsProviderURL() string {
 	return c.WsProviderURL
 }
